backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
curl_cffi==0.11.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.21
django-allauth==0.61.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.14.0
exceptiongroup==1.3.0
executing==2.2.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.18.1
jedi==0.19.2
jiter==0.10.0
matplotlib-inline==0.1.7
multitasking==0.0.11
nose==1.3.7
numpy==2.0.2
oauthlib==3.2.2
openai==1.82.0
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
peewee==3.18.1
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
websockets==15.0.1
yfinance==0.2.61
